+++ concordium-sdk-examples/src/main/java/com/concordium/sdk/examples/GetModuleList.java
@@ -1,0 +1,49 @@
+package com.concordium.sdk.examples;
+
+import com.concordium.sdk.ClientV2;
+import com.concordium.sdk.Connection;
+import com.concordium.sdk.Credentials;
+import com.concordium.sdk.exceptions.BlockNotFoundException;
+import com.concordium.sdk.exceptions.ClientInitializationException;
+import com.concordium.sdk.requests.BlockHashInput;
+import picocli.CommandLine;
+import picocli.CommandLine.Command;
+import picocli.CommandLine.Option;
+
+import java.net.MalformedURLException;
+import java.net.URL;
+import java.util.concurrent.Callable;
+
+/**
+ * Gets Contract Modules ta {@link BlockHashInput#BEST} and prints them to console.
+ */
+@Command(name = "GetModuleList", mixinStandardHelpOptions = true)
+public class GetModuleList implements Callable<Integer> {
+    @Option(
+            names = {"--endpoint"},
+            description = "GRPC interface of the node.",
+            defaultValue = "http://localhost:20000")
+    private String endpoint;
+
+    @Override
+    public Integer call() throws ClientInitializationException, MalformedURLException, BlockNotFoundException {
+        URL endpointUrl = new URL(this.endpoint);
+        Connection connection = Connection.builder()
+                .host(endpointUrl.getHost())
+                .port(endpointUrl.getPort())
+                .credentials(new Credentials())
+                .build();
+
+        ClientV2
+                .from(connection)
+                .getModuleList(BlockHashInput.BEST)
+                .forEachRemaining(System.out::println);
+
+        return 0;
+    }
+
+    public static void main(String[] args) {
+        int exitCode = new CommandLine(new GetModuleList()).execute(args);
+        System.exit(exitCode);
+    }
+}