+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 - Added support for GRPC V2 `GetPoolDelegators` for getting the delegators for a specified baking pool at a certain block.
 - Added support for GRPC V2 `GetElectionInfo` for getting the bakers at a certain block and other metadata of the election.
 - Added support for GRPC V2 `GetPoolInfo` for gettting information about a given pool at the end of a given block.
+- Added support for GRPC V2 `GetModuleList` for getting list of smart contract module at the end of a given block.
 
 ## 4.2.0
 - Added initial support for GRPC V2